+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 9.11.17
 
-### commit
+### commit 365a86ef5eacf5c75c71f5dfb13938c41ccdcb99
 
 #### Cop Drone
 * added ReportCrime()